+++ src/Storefront/Controller/StyleguideController.php
@@ -1,20 +1,25 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace StrixNLUxUpgrades\Storefront\Controller;
 
-use Shopware\Storefront\Controller\StorefrontController;
 use Shopware\Core\PlatformRequest;
 use Shopware\Core\System\SystemConfig\SystemConfigService;
+use Shopware\Storefront\Controller\StorefrontController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Attribute\Route;
 
-#[Route(defaults: ['_routeScope' => ['storefront']])]
+#[Route(defaults: [
+    '_routeScope' => ['storefront'],
+])]
 class StyleguideController extends StorefrontController
 {
     public function __construct(
         private readonly SystemConfigService $systemConfig
-    ) {}
+    ) {
+    }
 
     #[Route(path: '/styleguide', name: 'frontend.page.styleguide', methods: ['GET'])]
     public function index(): Response
@@ -23,7 +28,7 @@ class StyleguideController extends StorefrontController
         $salesChannelId = $request?->attributes->get(PlatformRequest::ATTRIBUTE_SALES_CHANNEL_ID);
 
         $enabled = (bool) ($this->systemConfig->get('StrixNLUxUpgrades.config.styleguideEnabled', $salesChannelId) ?? false);
-        if (!$enabled) {
+        if (! $enabled) {
             throw new NotFoundHttpException();
         }